requirements.txt
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
WTForms==2.3.3
email_validator
flask-sqlalchemy
flask_mysqldb
pymysql
cryptography
flask_login